Old version of Stage component replaced with the new code.
Stage2 was a temporary name.
